Read write seperation (#463)
#167 KAP Plus read/write separation

+++ build/bin/check-hive-user.sh
@@ -2,6 +2,7 @@
 # Kyligence Inc. License
 
 source $(cd -P -- "$(dirname -- "$0")" && pwd -P)/header.sh
+source $(cd -P -- "$(dirname -- "$0")" && pwd -P)/find-hadoop-conf-dir.sh
 
 echo "Checking Hive write permission..."
 
@@ -17,6 +18,12 @@ RANDNAME=chkenv__${RANDOM}
 HIVE_TEST_TABLE=${HIVE_TEST_DB}.${RANDNAME}
 HIVE_TEST_TABLE_LOCATION=${WORKING_DIR}/${RANDNAME}
 
+if [ -z "${kylin_hadoop_conf_dir}" ]; then
+    hadoop_conf_param=
+else
+    hadoop_conf_param="--config ${kylin_hadoop_conf_dir}"
+fi
+
 if [ "${HIVE_CLIENT_TYPE}" = "cli" ] 
 then
     hive -e "drop table if exists ${HIVE_TEST_TABLE}; create external table ${HIVE_TEST_TABLE} (name STRING,age INT) ROW FORMAT DELIMITED FIELDS TERMINATED BY ',' STORED AS TEXTFILE location '$HIVE_TEST_TABLE_LOCATION'; insert into table ${HIVE_TEST_TABLE} values ('"kylin"',1);"
@@ -26,9 +33,6 @@ then
     export ENABLE_CHECK_ENV=false
     ${dir}/kylin.sh io.kyligence.kap.source.hive.tool.CheckHCatalogJob ${HIVE_TEST_DB} ${RANDNAME} /tmp/kylin/check_hcatalog
     [[ $? == 0 ]] || quit "ERROR: Cannot get Hive table data via HCatInputFormat"
-    
-    hive -e "drop table ${HIVE_TEST_TABLE};"
-    hadoop fs -rm -R -skipTrash ${HIVE_TEST_TABLE_LOCATION}
 elif [ ${HIVE_CLIENT_TYPE} = "beeline" ]
 then
     HQL_TMP_FILE=hql_tmp__${RANDOM}
@@ -51,16 +55,22 @@ then
     export ENABLE_CHECK_ENV=false
     ${dir}/kylin.sh io.kyligence.kap.source.hive.tool.CheckHCatalogJob ${HIVE_TEST_DB} ${RANDNAME} /tmp/kylin/check_hcatalog
     [[ $? == 0 ]] || quit "ERROR: Cannot get Hive table data via HCatInputFormat"
-    
-    beeline ${HIVE_BEELINE_PARAM} -e "drop table ${HIVE_TEST_TABLE};"
-    rm -f ${HQL_TMP_FILE}
-    hadoop fs -rm -R -skipTrash ${HIVE_TEST_TABLE_LOCATION}
 else
     quit "ERROR: Only support 'cli' or 'beeline' hive client"
 fi
 
 # safeguard cleanup
 verbose "Safeguard cleanup..."
-hive -e "use ${HIVE_TEST_DB}; show tables 'chkenv__*';" | xargs -I '{}' hive -e "use ${HIVE_TEST_DB}; drop table {};"
-hadoop fs -rm -R -skipTrash "${WORKING_DIR}/chkenv__*"
+if [ "${HIVE_CLIENT_TYPE}" = "cli" ]
+then
+    hive -e "use ${HIVE_TEST_DB}; show tables 'chkenv__*';" | xargs -I '{}' hive -e "use ${HIVE_TEST_DB}; drop table {};"
+elif [ ${HIVE_CLIENT_TYPE} = "beeline" ]
+then
+    echo "use ${HIVE_TEST_DB};" > ${HQL_TMP_FILE}
+    echo "show tables 'chkenv__*';" >> ${HQL_TMP_FILE}
+    beeline ${HIVE_BEELINE_PARAM} -f ${HQL_TMP_FILE} | grep "chkenv__[[:digit:]]\+" -o | xargs -I "{}" beeline ${HIVE_BEELINE_PARAM} -e "drop table ${HIVE_TEST_DB}.{}"
+    rm -f ${HQL_TMP_FILE}
+fi
+
+hadoop ${hadoop_conf_param} fs -rm -R -skipTrash "${WORKING_DIR}/chkenv__*"
 exit 0
